+++ examples/single_line.rs
@@ -42,7 +42,6 @@ fn setup(mut commands: Commands, assets: Res<AssetServer>) {
             TextInputPrompt::default(),
             Node {
                 width: Val::Px(250.),
-                height: Val::Px(25.),
                 ..default()
             },
             BackgroundColor(NAVY.into()),